backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
Pillow==10.4.0
ImageHash==4.3.1
python-multipart==0.0.9